chore(lint): rename internal/cnpi/plugin to integration
Rename the plugin package to integration to avoid conflicting with
the Go standard library "plugin" package name. This eliminates the
need for part of the revive linter exclusion rule.
