+++ internal/cnpi/integration/client/cluster_test.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/cloudnative-pg/cnpg-i/pkg/operator"
 
-	"github.com/cloudnative-pg/cloudnative-pg/internal/cnpi/plugin/connection"
+	"github.com/cloudnative-pg/cloudnative-pg/internal/cnpi/integration/connection"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
