+++ internal/cnpi/integration/client/errors.go
