+++ internal/cnpi/integration/client/errors_test.go
