+++ internal/cnpi/integration/client/lifecycle.go
@@ -34,8 +34,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 
-	"github.com/cloudnative-pg/cloudnative-pg/internal/cnpi/plugin"
-	"github.com/cloudnative-pg/cloudnative-pg/internal/cnpi/plugin/connection"
+	"github.com/cloudnative-pg/cloudnative-pg/internal/cnpi/integration"
+	"github.com/cloudnative-pg/cloudnative-pg/internal/cnpi/integration/connection"
 )
 
 var runtimeScheme = runtime.NewScheme()
@@ -46,7 +46,7 @@ func init() {
 
 func (data *data) LifecycleHook(
 	ctx context.Context,
-	operationType plugin.OperationVerb,
+	operationType integration.OperationVerb,
 	cluster client.Object,
 	object client.Object,
 ) (client.Object, error) {
@@ -56,7 +56,7 @@ func (data *data) LifecycleHook(
 
 func (data *data) innerLifecycleHook(
 	ctx context.Context,
-	operationType plugin.OperationVerb,
+	operationType integration.OperationVerb,
 	cluster client.Object,
 	object client.Object,
 ) (client.Object, error) {
