+++ internal/cnpi/integration/connection/connection.go
