+++ internal/cnpi/integration/connection/doc.go
