+++ internal/cnpi/integration/connection/metadata.go
