+++ internal/cnpi/integration/connection/remote.go
