+++ internal/cnpi/integration/mapping.go
@@ -17,7 +17,7 @@ limitations under the License.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package plugin
+package integration
 
 import (
 	"fmt"
