+++ internal/controller/cluster_plugins.go
@@ -27,7 +27,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
-	cnpgiclient "github.com/cloudnative-pg/cloudnative-pg/internal/cnpi/plugin/client"
+	cnpgiclient "github.com/cloudnative-pg/cloudnative-pg/internal/cnpi/integration/client"
 )
 
 // updatePluginsStatus ensures that we load the plugins that are required to reconcile
